Fixing cyclic dependency
