+++ models/constants.go
@@ -5,23 +5,3 @@ const (
 	ByteSize             = 8
 	WifiThreshold        = 0.75
 )
-
-type Language int
-
-const (
-	Portuguese Language = iota
-	English
-)
-
-func (l Language) String() string {
-	switch l {
-	case Portuguese:
-		return "Portuguese"
-
-	case English:
-		return "English"
-
-	default:
-		return "unknown"
-	}
-}
