+++ translation/baseTypes.go
@@ -2,14 +2,14 @@ package translation
 
 import (
 	"religare/config"
-	"religare/models"
+	"religare/customTypes"
 )
 
 type TextIndex int
 
 type Texts map[TextIndex]string
 
-type LanguageMap map[models.Language]Texts
+type LanguageMap map[customTypes.Language]Texts
 
 func (l LanguageMap) GetValue(index TextIndex) string {
 	return l[config.AppLanguage][index]
